payments/src/routes/new.ts
import {
    BadRequestError,
    NotAuthorizedError,
    NotFoundError,
    OrderStatus,
    requireAuth,
    validateRequest,
} from '@scticketscommon/common';
import express, { Request, Response } from 'express';
import { body } from 'express-validator';
import { Order } from '../models/order';
import { stripe } from '../stripe';
import { Payment } from '../models/payments';

const router = express.Router();

router.post(
    '/api/payments',
    requireAuth,
    [body('token').not().isEmpty(), body('orderId').not().isEmpty()],
    validateRequest,
    async (req: Request, res: Response) => {
        const { token, orderId } = req.body;

        const order = await Order.findById(orderId);

        if (!order) {
            throw new NotFoundError();
        }

        if (order.userId !== req.currentUser!.id) {
            throw new NotAuthorizedError();
        }

        if (order.status === OrderStatus.Cancelled) {
            throw new BadRequestError('Cannot pay for a cancelled order');
        }

        const paymentIntents = await stripe.paymentIntents.create({
            currency: 'inr',
            amount: order.price * 100,
            automatic_payment_methods: {
                enabled: true,
            },
        });

        const payment = Payment.build({
            orderId,
            stripeId: paymentIntents.id,
        });

        await payment.save();

        res.status(201).send({ success: true });
    }
);

export { router as createChargeRouter };
